Factor out common code into base module/class
This is science fiction for now, but there's a lot of code that can be
reused/shared between `BloomFilter` and `MemLock`.  Refactor this code
into the `base` module and `Base` class.
